Add configuration for json utf characters handling.
The default pythons json.dumps behavior escapes any special utf8
characters, leaving \uxxxx encoded strings as the output. It's
sufficient to only pass a boolean flag for it to output the actual
characters. Add new global config option in vim that will control this
aspect of request result presentation - leaving the previous behavior
as the default.